# api-docs.yaml
openapi: 3.0.0
info:
  title: NXAPI Shop API (Event Edition)
  description: NXAPI Shop backend API documentation converted to manage Events using Prisma, NeonDB, and Cloudinary.
  version: 1.0.0
servers:
  - url: /api
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        fullName:
          type: string
        isActive:
          type: boolean
        roles:
          type: array
          items:
            type: string
    Event:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        date:
          type: string
          format: date-time
        location:
          type: string
        price:
          type: number
        slug:
          type: string
        capacity:
          type: integer
        images:
          type: array
          items:
            type: string
        user:
          $ref: '#/components/schemas/User'
    EventImage:
      type: object
      properties:
        id:
          type: integer
        url:
          type: string
        eventId:
          type: string

security:
  - bearerAuth: []

paths:
  /auth/register:
    post:
      summary: Register a new user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - fullName
              properties:
                email:
                  type: string
                password:
                  type: string
                fullName:
                  type: string
      responses:
        '201':
          description: User registered successfully
        '400':
          description: Missing required fields
        '409':
          description: User already exists
  
  /auth/login:
    post:
      summary: Login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '201':
          description: Login successful
        '400':
          description: Missing email or password
        '401':
          description: Invalid password
        '404':
          description: User not found

  /events:
    get:
      summary: Get all events (paginated)
      security: []
      parameters:
        - in: query
          name: offset
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        '200':
          description: A list of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
        '500':
          description: Internal server error
    post:
      summary: Create a newly listed event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - description
                - date
                - location
                - price
                - slug
                - capacity
                - images
              properties:
                title:
                  type: string
                description:
                  type: string
                date:
                  type: string
                  format: date-time
                location:
                  type: string
                price:
                  type: number
                slug:
                  type: string
                capacity:
                  type: integer
                images:
                  type: array
                  items:
                    type: string
      responses:
        '201':
          description: Event created successfully
        '400':
          description: Missing required data
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '409':
          description: Event title or slug already exists
        '500':
          description: Internal server error
  
  /events/all/{term}:
    get:
      summary: Search events by term (title or slug)
      security: []
      parameters:
        - in: path
          name: term
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful search results
        '500':
          description: Internal server error

  /events/{id}:
    get:
      summary: Get an event by its UUID or friendly Slug
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Event ID or Slug
      responses:
        '200':
          description: The requested event details
        '404':
          description: Event not found
        '500':
          description: Internal server error
    put:
      summary: Update an event details
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The UUID of the event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                date:
                  type: string
                  format: date-time
                location:
                  type: string
                price:
                  type: number
                slug:
                  type: string
                capacity:
                  type: integer
                images:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Event updated successfully
        '401':
          description: Unauthorized
        '404':
          description: Event not found
        '500':
          description: Internal server error
    delete:
      summary: Delete an event
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted successfully
        '401':
          description: Unauthorized
        '404':
          description: Event not found
        '409':
          description: Constraint conflict
        '500':
          description: Internal server error

  /events/{id}/images:
    post:
      summary: Upload an image to Cloudinary and attach to Event
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        '201':
          description: Image uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventImage'
        '400':
          description: Image is required
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '500':
          description: Internal server error

  /events/{id}/images/{imageId}:
    delete:
      summary: Delete an EventImage from DB and Cloudinary
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: path
          name: imageId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Image deleted successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Image not found
        '500':
          description: Internal server error